requirements.txt
scrapy==1.1
IMDbPY==5.0
sendgrid-python==2.2.1
